backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
gunicorn==21.2.0
supabase==2.0.2
pydantic[email]==2.5.0
python-dotenv==1.0.0
python-multipart==0.0.6
email-validator==2.1.0
requests==2.31.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
